Converts repo to ESNext
