+++ src/constants/index.js
@@ -1,0 +1,8 @@
+"use strict"
+
+export const statuses = [ "DENY", "ALLOW", "DISMISS" ]
+export const statusDeny = "DENY"
+export const statusAllow = "ALLOW"
+export const statusDismiss = "DISMISS"
+
+export const categories = [ "UNCATEGORIZED", "ESSENTIAL", "PERSONALIZATION", "ANALYTICS", "MARKETING" ]
